+++ crates/holochain/tests/tests/countersigning.rs
@@ -1,6 +1,4 @@
-use hdk::prelude::{
-    PreflightRequest, PreflightRequestAcceptance, Signature, SignedActionHashed, Timestamp,
-};
+use hdk::prelude::{PreflightRequest, PreflightRequestAcceptance};
 use holo_hash::{ActionHash, EntryHash};
 use holochain::conductor::api::error::{ConductorApiError, ConductorApiResult};
 use holochain::conductor::CellError;
@@ -9,7 +7,6 @@ use holochain::prelude::CountersigningSessionState;
 use holochain::sweettest::{
     await_consistency, SweetConductorBatch, SweetConductorConfig, SweetDnaFile,
 };
-use holochain_nonce::fresh_nonce;
 use holochain_state::prelude::{IncompleteCommitReason, SourceChainError};
 use holochain_types::app::DisabledAppReason;
 use holochain_types::prelude::Signal;
@@ -17,7 +14,6 @@ use holochain_types::signal::SystemSignal;
 use holochain_wasm_test_utils::TestWasm;
 use holochain_zome_types::countersigning::Role;
 use matches::assert_matches;
-use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::broadcast::Receiver;
 
@@ -759,6 +755,13 @@ async fn session_rollback_with_chc_enabled() {
     await_consistency(60, [alice, bob]).await.unwrap();
 }
 
+// This test is flaky on macos (locally?).
+// The problem only occurs when DPKI is enabled. The app is slow to react on the websocket, and therefore
+// the ws is closed and kept closed for 5 mins before it's re-connected.
+// As this is exclusive to DPKI, it's got something to do with DPKI execution, perhaps init or wasm
+// compilation, and is unrelated to websockets or kitsune.
+// When work on DPKI is resumed and this test is still flaky, this could serve as starting point for further
+// investigation.
 #[cfg(feature = "chc")]
 #[tokio::test(flavor = "multi_thread")]
 #[cfg_attr(
@@ -972,6 +975,7 @@ async fn multiple_agents_on_same_conductor_with_chc_enabled() {
 
 #[tokio::test(flavor = "multi_thread")]
 #[cfg_attr(target_os = "windows", ignore = "flaky")]
+#[cfg(feature = "chc")]
 async fn chc_should_respect_chain_lock() {
     holochain_trace::test_run();
 
@@ -1538,8 +1542,11 @@ async fn get_all_records(
     chc: &holochain_chc::ChcImpl,
 ) -> holochain_chc::ChcResult<
     Vec<(
-        SignedActionHashed,
-        Option<(Arc<holochain_chc::EncryptedEntry>, Signature)>,
+        hdk::prelude::SignedActionHashed,
+        Option<(
+            std::sync::Arc<holochain_chc::EncryptedEntry>,
+            hdk::prelude::Signature,
+        )>,
     )>,
 > {
     use holochain_types::fixt::SignatureFixturator;
@@ -1547,7 +1554,9 @@ async fn get_all_records(
     chc.get_record_data_request(holochain_chc::GetRecordsRequest {
         payload: holochain_chc::GetRecordsPayload {
             since_hash: None,
-            nonce: fresh_nonce(Timestamp::now()).unwrap().0,
+            nonce: holochain_nonce::fresh_nonce(hdk::prelude::Timestamp::now())
+                .unwrap()
+                .0,
         },
         signature: fixt::fixt!(Signature),
     })