+++ crates/holochain_types/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- InstalledAppCommon field `role_assignments` is now an IndexMap to ensure consisent ordering.
+- ZomeManifest dylib has been **deprecated**. Please use the wasm interpreter instead.
+
+## 0.5.0-dev.2
+
 ## 0.5.0-dev.1
 
 - InstalledAppCommon now includes a field `installed_at` with the timestamp that the app was installed